+++ package.json
@@ -20,5 +20,6 @@
   "scripts": {
     "start": "browser-sync start --server --no-ui --files \"js/**/*.js\"",
     "lint": "eslint js/"
-  }
+  },
+  "type": "module"
 }
